requirements-production.txt
# FASE 3C: Infraestrutura Avançada - Configuração Celery
# Instalar dependências para processamento assíncrono

# Background tasks
celery==5.4.0
redis==5.2.1

# Monitoramento
flower==2.0.1
django-celery-beat==2.8.0
django-celery-results==2.5.1

# Performance monitoring
django-debug-toolbar==5.1.0
django-extensions==3.2.3
sentry-sdk==2.20.0

# Database optimization
psycopg2-binary==2.9.11  # PostgreSQL production

# Cache
django-redis==5.4.0

# Rate limiting
django-ratelimit==4.1.0

# API documentation
drf-yasg==1.21.9

# Django REST framework
djangorestframework==3.15.1  # API REST

# Django Select2
django-select2==8.1.2  # Widgets avançados para formulários

gunicorn==21.2.0  # WSGI HTTP server para produção
python-decouple==3.8  # Leitura de variáveis de ambiente no Django
dj-database-url==1.3.0  # Configuração de banco via URL
django-widget-tweaks==1.4.12  # Customização de formulários nos templates
pandas==2.2.2  # Manipulação de dados em relatórios
openpyxl==3.1.5  # Manipulação de arquivos Excel
reportlab==4.3.1  # Geração de PDFs em relatórios
XlsxWriter==3.2.0  # Geração de arquivos Excel
